+++ src/pyjams/const/__init__.py
@@ -10,9 +10,11 @@ Subpackages
 .. autosummary::
    const
 """
-from .const import Pi, Pi2, Pi3, TwoPi, pi, pi2, pi3, Twopi, Sqrt2
-from .const import Gravity, T0, P0, T25, sigma, R, R_air, R_H2O
-from .const import Na, REarth, mmol_co2, mmol_h2o, mmol_air
+from .const import Pi, Pi2, Pi3, TwoPi, pi, pi2, pi3, Twopi, Sqrt2, sqrt2
+from .const import gravity, T0, P0, T25, sigma, R, Rair, Rh2o
+from .const import Na, kB, REarth
+from .const import mmol_co2, molmass_co2, mmol_h2o, molmass_h2o
+from .const import mmol_air, molmass_air
 from .const import density_quartz, cheat_quartz, cheat_water, cheat_air
 from .const import latentheat_vaporization
 from .const import R13VPDB, R18VSMOW, R2VSMOW
@@ -20,9 +22,10 @@ from .const import tiny, huge, eps
 
 
 __all__ = [
-    'Pi', 'Pi2', 'Pi3', 'TwoPi', 'pi', 'pi2', 'pi3', 'Twopi', 'Sqrt2',
-    'Gravity', 'T0', 'P0', 'T25', 'sigma', 'R', 'R_air', 'R_H2O',
-    'Na', 'REarth', 'mmol_co2', 'mmol_h2o', 'mmol_air',
+    'Pi', 'Pi2', 'Pi3', 'TwoPi', 'pi', 'pi2', 'pi3', 'Twopi', 'Sqrt2', 'sqrt2',
+    'gravity', 'T0', 'P0', 'T25', 'sigma', 'R', 'Rair', 'Rh2o',
+    'Na', 'kB', 'REarth', 'mmol_co2', 'molmass_co2', 'mmol_h2o', 'molmass_h2o',
+    'mmol_air', 'molmass_air',
     'density_quartz', 'cheat_quartz', 'cheat_water', 'cheat_air',
     'latentheat_vaporization',
     'R13VPDB', 'R18VSMOW', 'R2VSMOW',